requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.3
colorama==0.4.4
distlib==0.3.1
Django==3.1.7
django-import-export
filelock==3.0.12
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow
pylint==2.7.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.3
wrapt==1.12.1
gunicorn
